Refactor routing to use react-router-dom and add blank page for new internet speed
Rails handling routing was too much resonsibility, and a lot of that
overlaps with that reactjs gives us with their ecosystem, specifically
the npm package of `react-router-dom` which allows us to nagivate to
different pages and components without a full page reload.

This commit also adds a internet speeds blank page which we will add on
to next.