Add test.yml file for github continious integration
I struggled with getting rails' test environment to have the same
database name / username / password as the one that was spun up,
so I ended up hard coding the name/username/password in a separate
`github-actions.database.yml` file which replaces `database.yml`
only in github-actions. This results in the test environment recognizing
the database and being able to create the publicinternet test database
(and associated tables).

In the future, if we decide on github-actions as the CI of choice at the end,
we will have to DRY this up because it's not good practice to repeat variables,
we repeat this throughout the test.yml file.

Another thing to improve on, is because we have a separate database.yml file
just for CI, we may not completely match what the test environment represents, ideally
it's as close to production as possible to avoid us not catching something in CI and
only catching in production.